Completely re-did how Motors are handled
Deleted all the old Motor implementations and interfaces.

Created Motor, which is the replacement class for all the deleted Motor
classes we had.

Made Relay and MotorGroup to extend Motor.

Refactored all the classes referencing Motors/MotorInterface.

It all compiles and looks good, so good enough for me for tonight.